refactor(buildTrip): Exctract buildTrip to test utility
