+++ client/src/components/__tests__/TripList.test.tsx
@@ -2,35 +2,16 @@ import React from 'react';
 import {
   render,
   screen,
-  fireEvent,
   waitForElementToBeRemoved,
 } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import user from '@testing-library/user-event';
 import { server, rest } from '../../test/server/test-server';
-import faker from 'faker';
-import dayjs from 'dayjs';
 
-import { Trip, TripOverrides } from './../../types/Trip';
+import { buildTrip } from './../../test/utils/generate';
+
 import TripList from './../TripList';
 
 const apiUrl = process.env.REACT_APP_API_URL;
-
-function buildTrip(overrides: TripOverrides): Trip {
-  return {
-    _id: faker.random.uuid(),
-    name: faker.lorem.words(),
-    country: faker.address.country(),
-    startDate: transformTimestampToDate(faker.date.future()),
-    endDate: transformTimestampToDate(faker.date.future()),
-    stops: [faker.address.city(), faker.address.city()],
-    ...overrides,
-  };
-}
-
-function transformTimestampToDate(timestamp: Date): Date {
-  return new Date(dayjs().format('YYYY-MM-DD'));
-}
 
 test('displays a list of trips', async () => {
   const trips = [buildTrip({ name: 'Trip 1' }), buildTrip({ name: 'Trip 2' })];
